==> Project3/0_periodic_table_analysis.ipynb <==
import sys
sys.path.append("/Users/kisungkang/Documents/hackthon/hackathon_KRICT/periodic_trends") # Change the path

import json
import numpy as np
from collections import Counter
from periodic_trends import plotter

import pandas as pd
from tqdm import tqdm

df1 = pd.read_csv('./Data/1_MatDX/MatDX_nomad_EF_Binary.csv')
df2 = pd.read_csv('./Data/1_MatDX/MatDX_nomad_EF_Ternary.csv')

# Concatenate the DataFrames
df = pd.concat([df1, df2], ignore_index=True)

element_dict = {}

for index, row in tqdm(df.iterrows(), total=len(df), desc="Processing Rows"):
    for element in eval(row['formation_energy'])['reference'].keys():
        if element in element_dict.keys():
            element_dict[element] += 1
        else:
            element_dict[element] = 1

element_dict

df = pd.DataFrame(list(element_dict.items()), columns=['Element', 'Value'])

df.to_csv('element.csv', index=False)

plotter("element.csv")
