==> Project3/1_download_all_data_available.ipynb <==
import json
import requests
import csv
import pandas as pd 

type_of_compound = [
                    'Unary', 'Binary', 'Ternary', 'Quaternary',
                    'Quinary', 'Sexiary', 'Septenary', 'Octanary', 'Nonary', 'Decinary' 
                    ]

# Define base URL and endpoint
KRICTDB_URL = "https://materials.chemdx.org"
api_command = "/api/download/property"

# Set headers for the request
headers = {
    'Content-Type': 'application/json'
}
for compositional_arity in type_of_compound:
    # Define query parameters for the API request
    property_calculation = "Formation energy"

    # Construct the full API URL with query parameters
    api_url = f"{KRICTDB_URL}{api_command}?compositional_arity={compositional_arity}&property_calculation={property_calculation}"

    # Set a very large size to attempt to download all data at once
    option = json.dumps({
        "from": 1,
        "size": 10000  # Adjust size as needed if the API has a maximum limit per request
    })


    response = requests.post(api_url, data=option, headers=headers)

    if response.status_code == 200:
        data = response.json()
        print("Data fetched successfully! for ", compositional_arity)
        print("Total entries:", len(data))
        
        if len(data) > 0:
            
            df = pd.DataFrame(data)
            df.to_csv(f"./Data/MatDX_nomad_EF_{compositional_arity}.csv", index=False)
    else:
        print(f"Failed to fetch data. Status code: {response.status_code}")
