==> Project3/2_get_nomad_ref_orb-v2.ipynb <==
import pandas as pd
from tqdm import tqdm

df1 = pd.read_csv('./Data/1_MatDX/MatDX_nomad_EF_Binary.csv')
df2 = pd.read_csv('./Data/1_MatDX/MatDX_nomad_EF_Ternary.csv')

# Concatenate the DataFrames
df = pd.concat([df1, df2], ignore_index=True)

dict = {}

for index, row in tqdm(df.iterrows(), total=len(df), desc="Processing Rows"):
    for element in eval(row['formation_energy'])['reference'].keys():
        entry_id = eval(row['formation_energy'])['reference'][element].split('/')[-1]
        dict[element] = entry_id

print(dict)

import requests
import numpy as np
from ase import Atoms
# from mace.calculators import mace_mp
# calc = mace_mp(model="medium", dispersion=False, default_dtype='float64', device='cpu')
# from sevenn.sevennet_calculator import SevenNetCalculator
# sevennet_0_cal = SevenNetCalculator("7net-0", device='cpu', default_dtype='float64')  # 7net-0, SevenNet-0, 7net-0_22May2024, 7net-0_11July2024 ...
from orb_models.forcefield import pretrained
from orb_models.forcefield.calculator import ORBCalculator

device="cpu" # or device="cuda"
orbff = pretrained.orb_v2(device=device) # or choose another model using ORB_PRETRAINED_MODELS[model_name]()
calc = ORBCalculator(orbff, device=device)

E_ref = {}

for key in dict.keys():
    print(key)
    entry_id = dict[key]
    url = f"https://nomad-lab.eu/prod/rae/api/v1/entries/{entry_id}/archive"
    response = requests.get(url)
    if response.status_code == 200:
        data = response.json()
        
        # Navigate to the relevant structure data
        try:
            idealized_structure = data['data']['archive']['section_metadata']['encyclopedia']['material']['idealized_structure']
            
            # Extract necessary fields
            atom_labels = idealized_structure['atom_labels']
            atom_positions = idealized_structure['atom_positions']
            lattice_vectors = idealized_structure['lattice_vectors']

            # Convert to numpy arrays if not already in that format
            positions = np.array(atom_positions)
            cell = np.array(lattice_vectors) * 1e10

            # Create the ASE Atoms object
            refs = Atoms(symbols=atom_labels, scaled_positions=positions, cell=cell, pbc=True)
            # refs.calc = calc
            
            # E_mace = refs.get_potential_energy()

            refs.calc = calc

            E_m3gnet = refs.get_potential_energy()

            E_ref[key] = [len(refs), E_m3gnet]
        except KeyError as e:
            print(f"KeyError: Missing expected key in the data structure - {e}")

    else:
        print(f"Error: {response.status_code}")


E_ref

import json

# Specify the file name
file_name = "reference_orb-v2.json"

# Save the dictionary as a JSON file
with open(file_name, 'w') as json_file:
    json.dump(E_ref, json_file, indent=4)  # indent=4 makes the JSON pretty-printed

print(f"Dictionary saved as {file_name}")
