==> Project3/4_merge_descriptors.ipynb <==
import json
import numpy as np

# Specify the file name
file_name = "descriptors.json"

# Read the JSON file into a dictionary
with open(file_name, 'r') as json_file:
    ref_descr = json.load(json_file)

# Specify the file name
file_name = "descriptors_mace.json"

# Read the JSON file into a dictionary
with open(file_name, 'r') as json_file:
    ref_mace = json.load(json_file)

# Specify the file name
file_name = "descriptors_7net.json"

# Read the JSON file into a dictionary
with open(file_name, 'r') as json_file:
    ref_7net = json.load(json_file)

# Specify the file name
file_name = "descriptors_orb-v2.json"

# Read the JSON file into a dictionary
with open(file_name, 'r') as json_file:
    ref_orb_v2 = json.load(json_file)

data = {}

for key in ref_descr.keys():
    # Check if the key exists in all the reference dictionaries
    if key in ref_mace and key in ref_7net and key in ref_orb_v2:
        # Check if the last element of `ref_descr[key][0]` matches the first elements in the other references
        if (ref_descr[key][0][-1] == ref_mace[key][0][0] and 
            ref_descr[key][0][-1] == ref_7net[key][0][0] and 
            ref_descr[key][0][-1] == ref_orb_v2[key][0][0]):
            
            # Calculate the average of the values from the other references
            descr_new = ref_descr[key][0][:-1] + [
                np.average([ref_mace[key][0][1], ref_7net[key][0][1], ref_orb_v2[key][0][1]])
            ]
            forma_new = ref_descr[key][1]
            
            # Add the new entry to `data`
            data[key] = [descr_new, forma_new]
    # If the key is missing in any dictionary, it will automatically be skipped

print(f"The total number of data: {len(data)}")

# Specify the file name
output_file = 'descriptors_merged.json'

# Save the dictionary as a JSON file
with open(output_file, 'w') as json_file:
    json.dump(data, json_file, indent=4)  # indent=4 makes the JSON pretty-printed

print(f"Dictionary saved as {output_file}")
